+++ api/iMessage.go
@@ -3,7 +3,7 @@ package api
 // 消息封装抽象接口
 type IMessage interface {
 	// 获取消息 Id
-	GetMsgId() uint32
+	GetMsgId() int64
 
 	// 获取消息数据段长度
 	GetDataLen() uint32
@@ -12,7 +12,7 @@ type IMessage interface {
 	GetData() []byte
 
 	// 设置消息 Id
-	SetMsgId(uint32)
+	SetMsgId(int64)
 
 	// 设置消息内容
 	SetData([]byte)
